--- apps/c/sqlite3/sqlite3.patch
@@ -1,7 +1,0 @@
---- sqlite3.c	2023-03-22 00:38:57.962176976 +0800
-+++ /home/yufeng/Downloads/sqlite-amalgamation-3410100/sqlite3.c	2023-03-10 20:59:24.000000000 +0800
-@@ -1,4 +1,3 @@
--#include "todo.h"
- /******************************************************************************
- ** This file is an amalgamation of many separate C source files from SQLite
- ** version 3.41.1.  By combining all the individual C code files into this
