+++ ulib/c_libax/include/stdbool.h
@@ -3,9 +3,9 @@
 
 /* Represents true-or-false values */
 #ifndef __cplusplus
-#define true 1
+#define true  1
 #define false 0
-#define bool _Bool
+#define bool  _Bool
 #endif
 
 #endif // __STDBOOL_H__
