+++ ulib/c_libax/include/stdlib.h
@@ -12,5 +12,6 @@ void free(void *addr);
 #endif
 
 _Noreturn void abort(void);
+char *getenv(const char *name);
 
 #endif //__STDLIB_H__
