+++ ulib/c_libax/include/sys/time.h
@@ -6,4 +6,13 @@ struct timeval {
     long tv_usec; /* microseconds */
 };
 
+struct timezone {
+    int tz_minuteswest; /* (minutes west of Greenwich) */
+    int tz_dsttime;     /* (type of DST correction) */
+};
+
+typedef long timezone;
+int gettimeofday(struct timeval *tv, struct timezone *tz);
+int utimes(const char *filename, const struct timeval times[2]);
+
 #endif
