+++ ulib/c_libax/src/sys/stat.c
@@ -1,0 +1,16 @@
+#include <stdio.h>
+#include <sys/types.h>
+
+// TODO:
+int fchmod(int fd, mode_t mode)
+{
+    printf("%s%s\n", "Error: no ax_call implementation for ", __func__);
+    return 0;
+}
+
+// TODO:
+int mkdir(const char *pathname, mode_t mode)
+{
+    printf("%s%s\n", "Error: no ax_call implementation for ", __func__);
+    return 0;
+}
